Consolidate container tests into cli_test.go
Merged cli_container_test.go into cli_test.go to keep all CLI tests in
a single file for better organization and maintainability.

Changes:
- Moved TestFindDewySocketFiles to cli_test.go
- Moved TestDisplayContainerList to cli_test.go
- Moved TestGetAdminSocketPath to cli_test.go
- Added necessary imports (os, path/filepath, time, container package)
- Deleted cli_container_test.go

All tests pass with no changes to functionality.
